Refactors sync worker
