+++ frontend/package.json
@@ -6,11 +6,11 @@
     "build:css": "pnpm run generate:css -- --minify",
     "build:dev": "pnpm run prisma:generate",
     "build:remix": "remix build",
-    "build:worker": "tsc ./worker/sync.ts --outDir ./worker/build",
+    "build:worker": "tsc ./app/workers/sync.ts --outDir ./build",
     "dev": "run-p \"dev:*\"",
     "dev:css": "pnpm run generate:css --watch",
     "dev:remix": "remix dev",
-    "dev:worker": "pnpm run build:worker && node ./worker/build/worker/sync.js",
+    "dev:worker": "pnpm run build:worker && node ./build/workers/sync.js",
     "lint:fix": "eslint . --fix",
     "lint": "eslint .",
     "prisma:generate": "prisma generate",
